Combine Category objects
